App/requirements.txt
requests>=2.25.1
rich>=10.0.0
python-dotenv>=0.19.0
